requirements.txt
uvicorn
jsonschema
line-bot-sdk
paho-mqtt
django-bootstrap4
django-crispy-forms
crispy-bootstrap5
openai
django-cors-headers
channels
mysqlclient
Pillow
qrcode[pil]